Move some files to sub-package.
