+++ launchdarkly/environment_utils.go
@@ -1,4 +1,4 @@
-package main
+package launchdarkly
 
 import "strconv"
