+++ launchdarkly/resource_project.go
@@ -1,4 +1,4 @@
-package main
+package launchdarkly
 
 import (
 	"github.com/hashicorp/terraform/helper/schema"
